requirements.txt
Django~=2.0.6
python-jose
social-auth-app-django
pydgraph
nltk
snowballstemmer
